Replace sji.py with new_sji.py and rename.
